project-2-ai-chatbot/requirements.txt

fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langchain-openai
langchain-community
langgraph
tavily-python
pydantic
pytest
httpx
sqlalchemy
psycopg2-binary
python-multipart
PyPDF2
sentence-transformers
